ARM: dts: msm: Add addr and size cells for 6.13 kernel
Update dtsi as per LTS 6.13 upgrade compatibility to add
address and size cell in current node. On previous kernel it
use to inherit the address and size cell from parent node.
But now due to missing address and size cell in current
node, it is leading to null pointer dereference.

Change-Id: I93182696edcc56c8e504d5bb53b6f3d4b53328d1
CRs-Fixed: 4216528